+++ packages/platforms/browser/lib/index.ts
@@ -1,4 +1,4 @@
-export { DefaultRoutingProvider, default, onSettle } from './browser'
+export { DefaultRoutingProvider, default, onSettle, ReactRouterRoutingProvider } from './browser'
 export { type BrowserConfiguration } from './config'
-export { type RouteResolver, type RoutingProvider } from './routing-provider'
+export { type RouteResolver, type RoutingProvider, type StartRouteChangeCallback } from './routing-provider'
 export { type SpanOptions, type Time } from '@bugsnag/core-performance'