+++ packages/platforms/browser/tests/auto-instrumentation/route-change-plugin.test.ts
@@ -6,7 +6,7 @@
 import { InMemoryDelivery, IncrementingClock, VALID_API_KEY, createTestClient } from '@bugsnag/js-performance-test-utilities'
 import { RouteChangePlugin } from '../../lib/auto-instrumentation/route-change-plugin'
 import { createSchema, type BrowserConfiguration, type BrowserSchema } from '../../lib/config'
-import { createDefaultRoutingProvider } from '../../lib/default-routing-provider'
+import { createDefaultRoutingProvider } from '../../lib/routing-provider/default-routing-provider'
 import { type OnSettle } from '../../lib/on-settle'
 import { type StartRouteChangeCallback } from '../../lib/routing-provider'