+++ test/browser/features/fixtures/packages/rollup.config.mjs
@@ -1,5 +1,7 @@
 import nodeResolve from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
+import babel from '@rollup/plugin-babel';
+import replace from '@rollup/plugin-replace';
 import path from 'path'
 import url from 'url'
 
@@ -24,8 +26,15 @@ export default {
     ...(isCdnBuild ? cdnOutputOptions : {}),
   },
   plugins: [
-    nodeResolve({ browser: true, jail: path.resolve(`${__dirname}/..`) }),
-    commonjs()
+    nodeResolve({ browser: true, jail: path.resolve(`${__dirname}/..`), extensions: ['.mjs', '.js', '.json', '.node', '.jsx'] }),
+    commonjs(),
+    babel({ babelHelpers: 'bundled' }),
+    replace({
+      preventAssignment: true,
+      values: {
+        'process.env.NODE_ENV': JSON.stringify('production'),
+      },
+    })
   ],
   ...(isCdnBuild ? { external: ['@bugsnag/browser-performance'] } : {}),
   onLog (level, log, defaultHandler) {